Extract utils from controller
